test(factory-builder): split the greeting cells into their own file
Adding the greet-or-build cells pushed invocation_test.go past the 1000-line
file limit. The greeting behavior is a distinct concern from the build-and-
install flow the original file covers, so it gets its own file rather than an
exemption.
